float cast

+++ utils.py
@@ -48,8 +48,8 @@ def cross_entropy_loss():
     return fn
          
 def dice_coef_a(y_true, y_pred, smooth=100):
-    y_true_f = K.flatten(y_true)
-    y_pred_f = K.flatten(y_pred)
+    y_true_f = K.flatten(K.cast(y_true, tf.float32))
+    y_pred_f = K.flatten(K.cast(y_pred, tf.float32))
     return - 2 * (mime_U(y_true_f, y_pred_f)  - mime_I(y_true_f, y_pred_f)) / (np.square(mime_U(y_true_f, y_pred_f)))
 
 def dice_coef_b(y_true, y_pred, smooth=100):
